ci: repair cicd pipeline (#195)
Decompose backend pipeline into validate(ci) and release(cd) stages.
Run validate pipeline when a PR is opened and syncronized. When the
PR is merged bump version in package*.json automatically, commit,
tag the commit, build the image, tag the image, publish and redeploy.

Before the change the pipelines were in one file and needed
conditions to prevent steps from running on every change. Also the
tagging was mistakenly performed on the latest commit in the PR and
not in the master. So when the PR was merged the version tag stayed
in the last PR commit and had to be moved manually. Now no more
manual tagging needs to be performed, it happens automatically
when conventional commits (https://www.conventionalcommits.org)
keywords are found.

Issue-Ref: #194