refactor: migrate to hexagonal arch directory structure
The new structure contains 4 main areas: core, adapters, common
and main.

Core contains the application business logic with
features, models, services and now also repositories. I moved
the repos into core because they are tightly coupled to it and
implement part of the logic on the persistence layer.

There are two kinds of adapters: driven and driving. Driven
are the db client and the http client i.e. dependencies which
are injected into the core and used by it. Driving are the
rest api and the runner i.e. clients which are using the core
api.

Common contains utils which are used everywhere.

Main brings everything together and starts the app.

Documentation was added.

A small issue with the local-db Docker Compose setup was
resolved and a delete local-db data script was added.

Issue-Ref: #196