fix: remove local diff from steam apps collection
I added a new aggregation method which does not perform any diff
calculations on recieved and stored steam apps. Instead the database
is taking care of that. It stores only new steam apps ignoring the
existing ones.

This should lower the memory consumption and prevent the application
from crashing on Render where its running at the time of this
writing.

Issue-Ref: #32