fix: move diff operation to db layer
Added a new aggregation method which does not perform any diff
calculations on recieved and stored steam apps. Instead the database
is taking care of that. It stores only new steam apps ignoring the
existing ones.

This should lower the memory consumption and prevent the application
from crashing on Render where its running at the time of this
writing.

The method is collectSteamAppsDiffOnDbLayer and is used via a
simple environment variable feature toggle.

Changed the tests for the old and the new method to 
given-when-then style in descriptions and removed tests which knew
to much about the implementation details.

By mistake cicd-backend.yml was in master but it is deprecated and
was removed in #191. Removed it again.

The cloud db steam_apps collection had many duplications which are
cleaned up now. Also a database management documentation was
added.

Issue-Ref: #32