feature: removed updateGamesWithoutDetails method
# removed games repository method getGamesWithoutDetails
# removed games repository method updateGameDetailsFrom
# removed games aggregate method updateGameDetailsFrom
# removed game method updateGameDetailsFrom and all related code
# removed related tests

Previously, the application used steamDB to gather details for a game that we were not able to get from steam web, due to age restrictions or errors in steam web html page calls. I removed this functionality completely, as info on these games will be received through a different API, which will be added next.

Issue-Ref: #217