Feature/NR-69 Add checking for games via steam api (#222)
feature: added checking for games via steam api

# added checkIfGameViaSteamApi to game identifier
# added getSteamApiUntriedFilteredSteamApps method to steam apps repository
# added getSteamAppViaSteamApi method to steam client
# added recordSteamApiAttempt and updateAppTypeViaSteamAPi methods to steam app data model
# added fromSteamApi static method to game data model
# added steamApi to valid data sources
# added steam api responses mock
# added steam.api.responses.mock.js to prettier ignore

Previously the application was only using steam web to check for games. I added the steam api as a possible check for games that are either restricted or came across an error trying to check the game via steam web. The reason to use this api as a secondary option is potential lack of support of the api from Steam in the future and call limits.

Issue-Ref: #219